+++ src/test/java/cart/service/CartServiceTest.java
@@ -9,10 +9,13 @@ import static org.mockito.Mockito.when;
 
 import cart.exception.ErrorCode;
 import cart.exception.GlobalException;
+import cart.persistence.dao.CartDao;
+import cart.persistence.dao.MemberDao;
 import cart.persistence.entity.MemberCartEntity;
-import cart.persistence.repository.MemberCartRepository;
+import cart.persistence.entity.MemberEntity;
 import cart.service.dto.ProductResponse;
 import java.util.List;
+import java.util.Optional;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -26,7 +29,10 @@ import org.mockito.junit.jupiter.MockitoExtension;
 class CartServiceTest {
 
     @Mock
-    private MemberCartRepository memberProductRepository;
+    private CartDao cartDao;
+
+    @Mock
+    private MemberDao memberDao;
 
     @InjectMocks
     private CartService cartService;
@@ -35,7 +41,10 @@ class CartServiceTest {
     @DisplayName("특정 사용자의 장바구니에 물건을 추가한다.")
     void addCart() {
         // given
-        when(memberProductRepository.save(any(), any())).thenReturn(1L);
+        final MemberEntity memberEntity = new MemberEntity(1L, "USER",
+            "journey@gmail.com", "cGFzc3dvcmQ=", "져니", "010-1234-5678");
+        when(memberDao.findByEmail(any())).thenReturn(Optional.of(memberEntity));
+        when(cartDao.insert(any())).thenReturn(1L);
 
         // when
         long savedCartId = cartService.addCart("journey@gmail.com", 1L);
@@ -45,15 +54,47 @@ class CartServiceTest {
     }
 
     @Test
+    @DisplayName("장바구니 정보를 정상적으로 제거한다.")
+    void deleteCart_success() {
+        // given
+        final MemberEntity memberEntity = new MemberEntity(1L, "USER",
+            "journey@gmail.com", "cGFzc3dvcmQ=", "져니", "010-1234-5678");
+        when(memberDao.findByEmail(any())).thenReturn(Optional.of(memberEntity));
+        when(cartDao.deleteByMemberId(any(), any())).thenReturn(1);
+
+        // when, then
+        assertDoesNotThrow(() -> cartService.deleteCart("journey@gmail.com", 1L));
+    }
+
+    @ParameterizedTest(name = "장바구니 정보 삭제가 실패하면 예외가 발생한다.")
+    @ValueSource(ints = {0, 2})
+    void deleteCart_fail(final int deletedCount) {
+        // given
+        final MemberEntity memberEntity = new MemberEntity(1L, "USER",
+            "journey@gmail.com", "cGFzc3dvcmQ=", "져니", "010-1234-5678");
+        when(memberDao.findByEmail(any())).thenReturn(Optional.of(memberEntity));
+        when(cartDao.deleteByMemberId(any(), any())).thenReturn(deletedCount);
+
+        // when, then
+        assertThatThrownBy(() -> cartService.deleteCart("journey@gmail.com", 1L))
+            .isInstanceOf(GlobalException.class)
+            .extracting("errorCode")
+            .isEqualTo(ErrorCode.CART_INVALID_DELETE);
+    }
+
+    @Test
     @DisplayName("특정 사용자의 장바구니에 담긴 상품 리스트를 조회한다.")
     void getProductsByMemberEmail() {
         // given
+        final MemberEntity memberEntity = new MemberEntity(1L, "USER",
+            "journey@gmail.com", "cGFzc3dvcmQ=", "져니", "010-1234-5678");
+        when(memberDao.findByEmail(any())).thenReturn(Optional.of(memberEntity));
+
         final MemberCartEntity chickenProduct = new MemberCartEntity(1L, 1L, 1L,
             "치킨", "chicken_image_url", 20000, "KOREAN");
         final MemberCartEntity steakProductEntity = new MemberCartEntity(2L, 1L, 2L,
             "스테이크", "steakUrl", 40000, "WESTERN");
-        when(memberProductRepository.findByMemberEmail(any())).thenReturn(
-            List.of(chickenProduct, steakProductEntity));
+        when(cartDao.getProductsByMemberId(any())).thenReturn(List.of(chickenProduct, steakProductEntity));
 
         // when
         final List<ProductResponse> productResponses = cartService.getProductsByMemberEmail("journey@gmail.com");
@@ -65,29 +106,5 @@ class CartServiceTest {
             .containsExactly(
                 tuple(1L, "치킨", "chicken_image_url", 20000, "KOREAN"),
                 tuple(2L, "스테이크", "steakUrl", 40000, "WESTERN"));
-    }
-
-    @Test
-    @DisplayName("장바구니 정보를 정상적으로 제거한다.")
-    void deleteCart_success() {
-        // given
-        when(memberProductRepository.deleteByMemberEmail(any(), any())).thenReturn(1);
-
-        // when, then
-        assertDoesNotThrow(() -> cartService.deleteCart("journey@gmail.com", 1L));
-    }
-
-    @ParameterizedTest(name = "장바구니 정보 삭제가 실패하면 예외가 발생한다.")
-    @ValueSource(ints = {0, 2})
-    void deleteCart_fail(final int deletedCount) {
-        // given
-        when(memberProductRepository.deleteByMemberEmail(any(), any())).thenReturn(
-            deletedCount);
-
-        // when, then
-        assertThatThrownBy(() -> cartService.deleteCart("journey@gmail.com", 1L))
-            .isInstanceOf(GlobalException.class)
-            .extracting("errorCode")
-            .isEqualTo(ErrorCode.CART_INVALID_DELETE);
     }
 }
